requirements.txt
# AI New-Employee Onboarding Agent - Dependencies
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Database
psycopg2-binary>=2.9.9

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
